fix eval code.

+++ visualization/visualize_gen.py
@@ -28,8 +28,11 @@ def run_gen_until(
     output_dir = os.path.abspath(output_dir)    #转绝对路径
     os.makedirs(output_dir, exist_ok=True)
     overall_output_dir = os.path.join(output_dir, "overall")
-    if os.path.exists(overall_output_dir):
-        shutil.rmtree(overall_output_dir)
+    if file_save and vis_overall:
+        if os.path.exists(overall_output_dir):
+            shutil.rmtree(overall_output_dir)
+        os.makedirs(overall_output_dir, exist_ok=True)
+
 
     model = sampler.model
     tokenizer = sampler.tokenizer
@@ -38,6 +41,7 @@ def run_gen_until(
     key = 'LLaDABlock._manually_scaled_dot_product_attention'   #attention_map key
     steps_for_all = 0
     genlength_for_all = 0
+
     for i, prompt in enumerate(prompts, 1):
 
         get_local.cache[key] = []
@@ -54,13 +58,14 @@ def run_gen_until(
         exploration_intervals = OUT.exploration_intervals
         confidences = OUT.confidences
         transfer_idxs = OUT.transfer_idxs
+        metrics = OUT.metrics
 
-        print(f"prompt_{i} decoded over for {output_dir}/prompt_{i}")
-        actual_steps = len(outputs)
+        print(f"prompt_{i} decoded over, metrics={metrics}")
+        actual_steps = metrics.use_steps
         actual_genlength = len(outputs[0]) if len(outputs)>0 else 0
         steps_for_all += actual_steps
         genlength_for_all += actual_genlength
-        # 批量demask一下outputs
+
         outputs_decoded = decode_outputs(outputs, tokenizer)
         # 可视化阶段信息准备
         phase_records = None
@@ -69,13 +74,12 @@ def run_gen_until(
                 'phase_states': phase_states,
                 'exploration_intervals': exploration_intervals,
             }
-        n_prompt_tokens = len(input_ids)
+        n_prompt_tokens = input_ids.shape[1]
 
         # steps输出结果整体 可视化
         answer = tokenizer.batch_decode(out[:, input_ids.shape[1]:], skip_special_tokens=True)[0]
         if vis_overall:
-            if file_save:
-                os.makedirs(overall_output_dir, exist_ok=True)
+
             visualize_overall_steps(overall_output_dir, (outputs_decoded, confidences, transfer_idxs),
                 i, prompt, answer, is_show=console_show, is_save=file_save, phase_records=phase_records)
             print(f"overall saved to {overall_output_dir}")
@@ -83,40 +87,47 @@ def run_gen_until(
         # 逐 Step Attention Maps 可视化
         cache = get_local.cache
         if vis_attn_map:
+            existed_imgs_step = 0
+            attn_map_output_dir = ''
+            if file_save:
+                attn_map_output_dir = os.path.join(output_dir, f"Q{i}_details")
+                if os.path.exists(attn_map_output_dir):
+                    shutil.rmtree(attn_map_output_dir)
+                os.makedirs(attn_map_output_dir, exist_ok=True)
+
+                # existed_imgs_step = len(os.listdir(attn_map_output_dir))
+                # if existed_imgs_step == actual_steps:
+                #     print(f"prompt_{i} already done, go to next step")
+                #     continue
+                # elif existed_imgs_step > 0:
+                #     print(f"prompt_{i} partially done, continue generating from step{existed_imgs_step + 1}")
+                # else:
+                #     print(f"prompt_{i} fresh, starting generating its attention_maps")
+
+
             value_list = cache[key]
+            # print(f"---{len(value_list)}---")
             assert len(value_list) % actual_steps == 0
             n_layers = len(value_list) // actual_steps
+
             shape_per_layer = value_list[0].shape # (batch_size, heads, seq_len, seq_len)
-            # 重新塑形，将steps维度提取出来
             steps_attention_maps = np.array(value_list).reshape(
                 (actual_steps, n_layers) + shape_per_layer
             ).astype(np.float16) # (steps, N_layers, batch_size, heads, seq_len, seq_len)
+            # print("steps_attention_maps shape:", steps_attention_maps.shape)
+
             # --- 动态计算布局参数 ---
-            # 固定绘图的宽度
             # FIG_WIDTH = max(32, gen_length * 0.7)
-            FIG_WIDTH = 16
+            FIG_WIDTH = 32
             P1_HEIGHT = actual_steps * (FIG_WIDTH / actual_genlength) # decoding_history区域高度
             P2_HEIGHT = FIG_WIDTH * 3 / 4 # attention_maps区域高度
-            P3_HEIGHT = 1 # prompt&answer区域高度
+            P3_HEIGHT = 0 # prompt&answer区域高度
             FIG_HEIGHT = P1_HEIGHT + P2_HEIGHT
 
-            # 为当前prompt的每个step生成一张包含预测和所有层注意力的大图
-            prompt_output_dir = os.path.join(output_dir, f"prompt_{i}_details")
-            # if os.path.exists(prompt_output_dir):
-            #     shutil.rmtree(prompt_output_dir)
-            os.makedirs(prompt_output_dir, exist_ok=True)
-            existed_imgs_step = len(os.listdir(prompt_output_dir))
-            if existed_imgs_step == actual_steps:
-                print(f"prompt_{i} already done, go to next step")
-                continue
-            elif existed_imgs_step > 0:
-                print(f"prompt_{i} partially done, continue generating from step{existed_imgs_step + 1}")
-
+            # 绘制attention maps on each step
             for step_idx in range(existed_imgs_step, actual_steps):
-                # 创建一个足够大的画布
                 fig = plt.figure(figsize=(FIG_WIDTH, FIG_HEIGHT))
-
-                # 使用GridSpec创建 3x1 的布局，分别对应P1(decoding_history)和P2(attention_maps)
+                # 由上至下共3个区域
                 gs_main = gridspec.GridSpec(3, 1, figure=fig, height_ratios=[P1_HEIGHT, P2_HEIGHT, P3_HEIGHT], hspace=0.2)
 
                 # --- 1. 绘制上方的decoding_history大图 ---
@@ -131,11 +142,10 @@ def run_gen_until(
                 )
 
                 # --- 2. 绘制下方的attention_maps ---
-                gs_p2 = gridspec.GridSpecFromSubplotSpec(6, 8, subplot_spec=gs_main[1, 0], hspace=0.2)
+                gs_p2 = gridspec.GridSpecFromSubplotSpec(6, 6, subplot_spec=gs_main[1, 0], hspace=0.2)
                 # 中间画一张大的总平均attention_map。在N_layers和heads两个维度上取平均
-                ax_avg_all = fig.add_subplot(gs_p2[0:6, 1:7])
+                ax_avg_all = fig.add_subplot(gs_p2[:, :])
                 attn_data_all_avg = steps_attention_maps[step_idx, :, 0].mean(axis=(0, 1)) # shape: (seq, seq)
-
                 plot_single_attention_map_on_ax(
                     ax=ax_avg_all,
                     attention_map_data=attn_data_all_avg,
@@ -144,61 +154,41 @@ def run_gen_until(
                     transfers=transfer_idxs[step_idx]
                 )
 
-                # --- 3. 在最底部的区域专门放文字信息 ---
-                gs_p3 = gridspec.GridSpecFromSubplotSpec(1, 2, subplot_spec=gs_main[2, 0], wspace=0.1, hspace=0.1)
-                # prompt区域
-                ax_prompt = fig.add_subplot(gs_p3[0, 0]) # 占据第三行
-                ax_prompt.axis('off')
-                ax_prompt.text(
-                    0.5,  # x坐标 (0=最左, 1=最右)
-                    0.5,   # y坐标 (0=最下, 1=最上)
-                    f"Prompt: {prompt if len(prompt)<=2000 else prompt[:2000]+'...'}\n ",
-                    transform=ax_prompt.transAxes, # 使用相对于ax自身的坐标系，非常方便
-                    fontsize=20,
-                    horizontalalignment='center', # 新增：水平对齐方式设为居中
-                    verticalalignment='center',   # 修改：垂直对齐方式设为居中
-                )
-                # answer区域
-                ax_answer = fig.add_subplot(gs_p3[0, 1]) # 占据第三行
-                ax_answer.axis('off')
-                ax_answer.text(
-                    0.5,  # x坐标 (0=最左, 1=最右)
-                    0.5,   # y坐标 (0=最下, 1=最上)
-                    f"Answer: {answer if len(answer)<=2000 else answer[:2000]+'...'}\n ",
-                    transform=ax_answer.transAxes, # 使用相对于ax自身的坐标系，非常方便
-                    fontsize=20,
-                    horizontalalignment='center', # 新增：水平对齐方式设为居中
-                    verticalalignment='center',   # 修改：垂直对齐方式设为居中
-                )
+                plt.suptitle(f"******Answer******: {answer if len(answer) <= 2000 else answer[:2000]}", fontsize=16)
 
-                # 设置总标题
-                fig.suptitle(f"Detailed Analysis for Step {step_idx + 1}/{actual_steps}\n", fontsize=24)
-                # 保存图像，为每个prompt创建一个子文件夹
                 if file_save:
-                    save_path = os.path.join(prompt_output_dir, f"step_{step_idx + 1}.png")
+                    save_path = os.path.join(attn_map_output_dir, f"step_{step_idx + 1}.png")
                     plt.savefig(save_path)
-                    print(f"attention_maps saved to {prompt_output_dir}")
+                    # print(f"attention_map saved to {save_path}")
                 if console_show:
                     plt.show()
 
                 plt.close(fig) # 必须关闭，否则会在内存中累积
 
+            if file_save:
+                print(f"prompt{i}'s attention_maps have been saved to {attn_map_output_dir}")
+
         # 清空已处理过的attention_map，为下一个prompt做准备
         if key in cache:
-          cache[key] = []
+            cache[key] = []
+
     print(f"{len(prompts)} promtps over. avg steps reduced: {steps_for_all / len(prompts):.2f}X")
 
 
 if __name__ == "__main__":
     print(f"connected to server, current path: {os.path.abspath(__file__)}")
-    device = 'cuda:0'
+    device = 'cuda:1'
 
-    # get_local.activate()  # 在引入模型之前，激活装饰器
     # 提示词替换，与4-shot的行为保持一致
     few_shot_filename = "../prompts/gsm8k_shot.txt"
     with open(few_shot_filename, "r", encoding="utf-8") as f:
-        gsm8k_shot_prompts= f.readlines()
+        gsm8k_prompts= f.readlines()
 
+    # gsm8k_dataset = load_dataset('openai/gsm8k', 'main')
+    # gsm8k_prompts = gsm8k_dataset['test']['question'][0:5]
+
+
+    # get_local.activate()  # 在引入模型之前，激活装饰器
     model_path = "../models/LLaDA-8B-Instruct"
     config = MRSamplerConfig(
         cfg_scale=0.0,
@@ -206,41 +196,36 @@ if __name__ == "__main__":
         max_exploration_steps=10,
         exploration_N=2,
         exploration_M=3,
-        exploration_threshold=0.05,
-        acceleration_threshold=0.8,
+        exploration_threshold=0.25,
+        acceleration_parallel_method='fixed',
+        acceleration_threshold=0.9,
         acceleration_low_threshold=0.6,
         acceleration_factor=0.6,
         max_mopup_steps=10,
         mopup_gate_ratio=0.9,
-        mopup_speed=2
+        mopup_speed=2,
+        positional_weights_type='none',
+        max_weight=1.0,
+        initial_min_weight=0.1,
     )
     # 用户传入的kwargs覆盖默认参数
     sampler = MRSampler.from_path(model_path, config=config, device=device, torch_dtype=torch.bfloat16)
 
-    # max_exploration_steps_list = [3, 10, 20]
-    # for max_exploration_steps in max_exploration_steps_list:
-    #     sampler.max_exploration_steps = max_exploration_steps
-    #     # 测跑
-    #     run_gen_until(
-    #         sampler=sampler,
-    #         prompts=gsm8k_shot_prompts[:5],
-    #         max_steps=256,
-    #         gen_length=256,
-    #         output_dir=f"./imgs/Test/gsm8k_s256/N2E{max_exploration_steps}_min0.25",
-    #         console_show=False, file_save=True, vis_overall=True, vis_attn_map=False
-    #     )
-
     sampler.max_exploration_steps = 10
     sampler.exploration_N = 2
-    exploration_thresholds = [0.05, 0.15, 0.3, 0.5]
+    # exploration_thresholds = [0.05, 0.15, 0.3, 0.5]
+    # exploration_thresholds = [0.2, 0.25]  -> No Positionals Weights下, 0.25表现最好
+    exploration_thresholds = [0.25]
     for et in exploration_thresholds:
         sampler.exploration_threshold = et
+        # output_dir = f"./imgs/Test/gsm8k_s256/N{sampler.exploration_N}E{sampler.max_exploration_steps}_ET{et}_DPimin{sampler.initial_min_weight}_V1"
+        output_dir = f"./imgs/Test/gsm8k_s256/N{sampler.exploration_N}E{sampler.max_exploration_steps}_ET{et}_APM{sampler.acceleration_parallel_method}_V1"
         run_gen_until(
             sampler=sampler,
-            prompts=gsm8k_shot_prompts[:5],
+            prompts=gsm8k_prompts[:3],
             max_steps=256,
             gen_length=256,
-            output_dir=f"./imgs/Test/gsm8k_s256/N{sampler.exploration_N}E{sampler.max_exploration_steps}_ET{et}_min0.25",
+            output_dir=output_dir,
             console_show=False, file_save=True, vis_overall=True, vis_attn_map=False
         )
